+++ CHANGES.md
@@ -9,6 +9,7 @@ number or not.
 ## Gameplay and UI
 
 - Massive performance improvement, like an order of magnitude (see technical details for more)
+- Table filtering, like "show only players with >1 block per game and >1 steal per game"
 
 ## Technical details